terminado calculo de energia

+++ src/main/java/imoveis/aluguel/dtos/EnergyDtoRequest.java
@@ -2,7 +2,7 @@ package imoveis.aluguel.dtos;
 
 import java.time.LocalDate;
 
-public record EnergyDtoRequest(Long id, Long counter1, Long counter2, Long counter3, Double kwhValue, Double billAmount,
+public record EnergyDtoRequest(Long id, Double counter1, Double counter2, Double counter3, Double kwhValue, Double billAmount,
         LocalDate date) {
 
 }
