+++ src/main/java/imoveis/aluguel/dtos/EnergyDtoResponseList.java
@@ -2,7 +2,7 @@ package imoveis.aluguel.dtos;
 
 import java.time.LocalDate;
 
-public record EnergyDtoResponseList(Long id, Long counter1, Long counter2, Long counter3, Double amount1,
+public record EnergyDtoResponseList(Long id, Double counter1, Double counter2, Double counter3, Double amount1,
         Double amount2, Double amount3, Double kwhValue, Double billAmount, LocalDate date, Boolean last) {
 
 }
